% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infograf.R
\docType{package}
\name{infograf}
\alias{infograf}
\alias{infograf-package}
\title{Various functions to help with graph and information theory analyses/visualizations/etc.}
\description{
Various functions to help with graph and information theory analyses/visualizations/etc.
}
