% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDistMatrix.R
\name{MakeDistMatrix}
\alias{MakeDistMatrix}
\title{Distance matrix from data.table}
\usage{
MakeDistMatrix(dt, id.x, id.y, value.cn = "distance")
}
\arguments{
\item{dt}{a data.table}

\item{id.x}{String, name of identifier column x}

\item{id.y}{String, name of identifier column y}

\item{value.cn}{name of column with distances}
}
\value{
`dm`  Named distance `matrix`.
}
\description{
Makes a named distance matrix from a data.table of (x, y, distance(x, y)).
}
