% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeNonRedundantPairs.R
\name{MakeNonRedundantPairs}
\alias{MakeNonRedundantPairs}
\title{Make list of non-redundant (non-reciprocal) pairings}
\usage{
MakeNonRedundantPairs(things, prefix = "thing", suffix = c("i", "j"),
  sep = ".")
}
\arguments{
\item{things}{vector of things}

\item{prefix}{string, prefix of column names (e.g. "thing")}

\item{suffix}{character length 2, suffix of column names (usually c("i", "j"), e.g.)}

\item{sep}{separator joining prefix and suffix (e.g. ".")}
}
\value{
data.table with thing i in one column and thing j in another
}
\description{
Given list of things, return data.table of non-redundant pairs between
things (including self-self).
}
